+++ R/agepyramid.h.R
@@ -355,7 +355,7 @@ agepyramidBase <- if (requireNamespace("jmvcore", quietly=TRUE)) R6::R6Class(
             super$initialize(
                 package = "ClinicoPath",
                 name = "agepyramid",
-                version = c(1,0,7),
+                version = c(1,0,8),
                 options = options,
                 results = agepyramidResults$new(options=options),
                 data = data,