+++ R/alluvial.h.R
@@ -386,7 +386,7 @@ alluvialBase <- if (requireNamespace("jmvcore", quietly=TRUE)) R6::R6Class(
             super$initialize(
                 package = "ClinicoPath",
                 name = "alluvial",
-                version = c(1,0,7),
+                version = c(1,0,8),
                 options = options,
                 results = alluvialResults$new(options=options),
                 data = data,