+++ R/categorize.h.R
@@ -344,7 +344,7 @@ categorizeBase <- if (requireNamespace("jmvcore", quietly=TRUE)) R6::R6Class(
             super$initialize(
                 package = "ClinicoPath",
                 name = "categorize",
-                version = c(1,0,7),
+                version = c(1,0,8),
                 options = options,
                 results = categorizeResults$new(options=options),
                 data = data,