+++ R/reportcat.h.R
@@ -126,7 +126,7 @@ reportcatBase <- if (requireNamespace("jmvcore", quietly=TRUE)) R6::R6Class(
             super$initialize(
                 package = "ClinicoPath",
                 name = "reportcat",
-                version = c(1,0,7),
+                version = c(1,0,8),
                 options = options,
                 results = reportcatResults$new(options=options),
                 data = data,