+++ R/tableone.h.R
@@ -193,7 +193,7 @@ tableoneBase <- if (requireNamespace("jmvcore", quietly=TRUE)) R6::R6Class(
             super$initialize(
                 package = "ClinicoPath",
                 name = "tableone",
-                version = c(1,0,9),
+                version = c(1,0,8),
                 options = options,
                 results = tableoneResults$new(options=options),
                 data = data,