+++ R/venn.h.R
@@ -739,7 +739,7 @@ vennBase <- if (requireNamespace("jmvcore", quietly=TRUE)) R6::R6Class(
             super$initialize(
                 package = "ClinicoPath",
                 name = "venn",
-                version = c(1,0,7),
+                version = c(1,0,8),
                 options = options,
                 results = vennResults$new(options=options),
                 data = data,